made a lot of changes

+++ Core/Src/rpi-display/gfxfont.h
@@ -15,8 +15,8 @@ typedef struct{
   uint8_t width;         ///< Bitmap dimensions in pixels
   uint8_t height;        ///< Bitmap dimensions in pixels
   uint8_t xAdvance;      ///< Distance to advance cursor (x axis)
-  int8_t xOffset;        ///< X dist from cursor pos to UL corner
-  int8_t yOffset;        ///< Y dist from cursor pos to UL corner
+  int16_t xOffset;        ///< X dist from cursor pos to UL corner
+  int16_t yOffset;        ///< Y dist from cursor pos to UL corner
 } GFXglyph;
 
 /// Data stored for FONT AS A WHOLE
